Updated code to have a new composable in order to avoid code repetition
